Renamed test packages for SemverIntegTest
Task: #26
