+++ japicmp-testbase/japicmp-test-v1/src/main/java/japicmp/test/semver/addedPublicDeprecation/SemverTestee010a.java
@@ -1,0 +1,8 @@
+package japicmp.test.semver.addedPublicDeprecation;
+
+public class SemverTestee010a {
+
+	public void methodAnnotationDeprecatedAdded() {
+
+	}
+}
