+++ japicmp-testbase/japicmp-test-v1/src/main/java/japicmp/test/semver/changesFromDefaultToPublic/SemverTestee010bAccessModiferChangesFromDefaultToPublic.java
@@ -1,0 +1,5 @@
+package japicmp.test.semver.changesFromDefaultToPublic;
+
+class SemverTestee010bAccessModiferChangesFromDefaultToPublic {
+
+}
