+++ japicmp-testbase/japicmp-test-v1/src/main/java/japicmp/test/semver/changesFromPublicToDefault/SemverTestee100bAccessModifierChangesFromPublicToDefault.java
@@ -1,0 +1,5 @@
+package japicmp.test.semver.changesFromPublicToDefault;
+
+public class SemverTestee100bAccessModifierChangesFromPublicToDefault {
+
+}
