+++ japicmp-testbase/japicmp-test-v1/src/main/java/japicmp/test/semver/changesMethodFromPublicToDefault/SemverTestee100c.java
@@ -1,0 +1,8 @@
+package japicmp.test.semver.changesMethodFromPublicToDefault;
+
+public class SemverTestee100c {
+
+	public void methodAccessModifierChangesFromPublicToDefault() {
+
+	}
+}
