+++ japicmp-testbase/japicmp-test-v1/src/main/java/japicmp/test/semver/implementationChange/SemverTestee001a.java
@@ -1,0 +1,12 @@
+package japicmp.test.semver.implementationChange;
+
+public class SemverTestee001a {
+
+	public void methodUnchanged() {
+
+	}
+
+	public int methodImplementationChanges(int a, int b) {
+		return a + b;
+	}
+}
