+++ japicmp-testbase/japicmp-test-v1/src/main/java/japicmp/test/semver/publicMethodParamChange/SemverTestee100a.java
@@ -1,0 +1,9 @@
+package japicmp.test.semver.publicMethodParamChange;
+
+public class SemverTestee100a {
+
+	public void methodAnnotationDeprecatedAddedAndParameterIntAdded() {
+
+	}
+
+}
