+++ japicmp-testbase/japicmp-test-v1/src/main/java/japicmp/test/semver/superclassWithField/SemverTestee100d.java
@@ -1,0 +1,5 @@
+package japicmp.test.semver.superclassWithField;
+
+public class SemverTestee100d extends SemverTestee100dSuperclass {
+
+}
