--- japicmp-testbase/japicmp-test-v1/src/main/java/japicmp/test/semver010/a/SemverTestee010a.java
@@ -1,8 +1,0 @@
-package japicmp.test.semver010.a;
-
-public class SemverTestee010a {
-
-	public void methodAnnotationDeprecatedAdded() {
-
-	}
-}
