--- japicmp-testbase/japicmp-test-v1/src/main/java/japicmp/test/semver010/b/SemverTestee010bAccessModiferChangesFromDefaultToPublic.java
@@ -1,5 +1,0 @@
-package japicmp.test.semver010.b;
-
-class SemverTestee010bAccessModiferChangesFromDefaultToPublic {
-
-}
