--- japicmp-testbase/japicmp-test-v1/src/main/java/japicmp/test/semver100/a/SemverTestee100a.java
@@ -1,20 +1,0 @@
-package japicmp.test.semver100.a;
-
-public class SemverTestee100a {
-
-	public void methodAnnotationDeprecatedAddedAndParameterIntAdded() {
-
-	}
-
-	void methodAnnotationDeprecatedAdded() {
-
-	}
-
-	protected void methodReturnValueChangesFromVoidToString() {
-
-	}
-
-	private void methodRemoved() {
-
-	}
-}
