--- japicmp-testbase/japicmp-test-v1/src/main/java/japicmp/test/semver100/b/SemverTestee100bAccessModifierChangesFromPublicToDefault.java
@@ -1,5 +1,0 @@
-package japicmp.test.semver100.b;
-
-public class SemverTestee100bAccessModifierChangesFromPublicToDefault {
-
-}
