--- japicmp-testbase/japicmp-test-v1/src/main/java/japicmp/test/semver100/c/SemverTestee100c.java
@@ -1,8 +1,0 @@
-package japicmp.test.semver100.c;
-
-public class SemverTestee100c {
-
-	public void methodAccessModifierChangesFromPublicToDefault() {
-
-	}
-}
