--- japicmp-testbase/japicmp-test-v1/src/main/java/japicmp/test/semver100/d/SemverTestee100dSuperclass.java
@@ -1,5 +1,0 @@
-package japicmp.test.semver100.d;
-
-public class SemverTestee100dSuperclass {
-	public int fieldInSuperclass;
-}
