+++ japicmp-testbase/japicmp-test-v2/src/main/java/japicmp/test/semver/addedPublicDeprecation/SemverTestee010a.java
@@ -1,0 +1,9 @@
+package japicmp.test.semver.addedPublicDeprecation;
+
+public class SemverTestee010a {
+
+	@Deprecated
+	public void methodAnnotationDeprecatedAdded() {
+
+	}
+}
