+++ japicmp-testbase/japicmp-test-v2/src/main/java/japicmp/test/semver/changesFromDefaultToPublic/SemverTestee010bAccessModiferChangesFromDefaultToPublic.java
@@ -1,0 +1,5 @@
+package japicmp.test.semver.changesFromDefaultToPublic;
+
+public class SemverTestee010bAccessModiferChangesFromDefaultToPublic {
+
+}
