+++ japicmp-testbase/japicmp-test-v2/src/main/java/japicmp/test/semver/changesFromPublicToDefault/SemverTestee100bAccessModifierChangesFromPublicToDefault.java
@@ -1,0 +1,5 @@
+package japicmp.test.semver.changesFromPublicToDefault;
+
+class SemverTestee100bAccessModifierChangesFromPublicToDefault {
+
+}
