--- japicmp-testbase/japicmp-test-v2/src/main/java/japicmp/test/semver001/a/SemverTestee001a.java
@@ -1,12 +1,0 @@
-package japicmp.test.semver001.a;
-
-public class SemverTestee001a {
-
-	public void methodUnchanged() {
-
-	}
-
-	public int methodImplementationChanges(int a, int b) {
-		return a - b;
-	}
-}
