--- japicmp-testbase/japicmp-test-v2/src/main/java/japicmp/test/semver010/a/SemverTestee010a.java
@@ -1,9 +1,0 @@
-package japicmp.test.semver010.a;
-
-public class SemverTestee010a {
-
-	@Deprecated
-	public void methodAnnotationDeprecatedAdded() {
-
-	}
-}
