--- japicmp-testbase/japicmp-test-v2/src/main/java/japicmp/test/semver100/a/SemverTestee100a.java
@@ -1,23 +1,0 @@
-package japicmp.test.semver100.a;
-
-public class SemverTestee100a {
-
-	@Deprecated
-	public void methodAnnotationDeprecatedAddedAndParameterIntAdded(int value) {
-
-	}
-
-	@Deprecated
-	void methodAnnotationDeprecatedAdded() {
-
-	}
-
-	@Deprecated
-	protected String methodReturnValueChangesFromVoidToString() {
-		return null;
-	}
-
-	private void methodAdded() {
-
-	}
-}
