--- japicmp-testbase/japicmp-test-v2/src/main/java/japicmp/test/semver100/d/SemverTestee100d.java
@@ -1,5 +1,0 @@
-package japicmp.test.semver100.d;
-
-public class SemverTestee100d extends SemverTestee100dSuperclass {
-	protected int fieldInSuperclass;
-}
