+++ japicmp-testbase/japicmp-test/src/test/java/japicmp/test/output/SemverOutIntegTest.java
@@ -13,74 +13,58 @@ import japicmp.model.AccessModifier;
 import japicmp.model.JApiClass;
 import japicmp.output.semver.SemverOut;
 import japicmp.test.util.Helper;
-import org.junit.Ignore;
 import org.junit.Test;
 
 public class SemverOutIntegTest {
 
+	public static final String PATCH = "0.0.1";
+	public static final String MINOR = "0.1.0";
+	public static final String MAJOR = "1.0.0";
+
 	@Test
 	public void testSemver001_implementation_of_method_changes() {
-		String lastPackage = "semver001.a";
-		String string = getSemverDiff(lastPackage);
-		assertEquals("0.0.1", string);
+		assertSemverDiffByPackage(PATCH, "implementationChange");
 	}
 
 	@Test
 	public void testSemver010_added_deprecated_annotation() {
-		String lastPackage = "semver010.a";
-		String string = getSemverDiff(lastPackage);
-		assertEquals("0.1.0", string);
+		assertSemverDiffByPackage(MINOR, "addedPublicDeprecation"); // TODO is this MINOR?
 	}
 
 	@Test
 	public void testSemver010_increase_visibility_of_class_from_default_to_public() {
-		String lastPackage = "semver010.b";
-		String string = getSemverDiff(lastPackage);
-		assertEquals("0.1.0", string);
+		assertSemverDiffByPackage(MINOR, "changesFromDefaultToPublic");
 	}
 
 	@Test
 	public void testSemver100_change_method() {
-		String lastPackage = "semver100.a";
-
-		String string = getSemverDiff(lastPackage);
-
-		assertEquals("1.0.0", string);
+		assertSemverDiffByPackage(MAJOR, "publicMethodParamChange");
 	}
 
 	@Test
 	public void testSemver100_reduce_class_visibility() {
-		String lastPackage = "semver100.b";
-
-		String string = getSemverDiff(lastPackage);
-
-		assertEquals("1.0.0", string);
+		assertSemverDiffByPackage(MAJOR, "changesFromPublicToDefault");
 	}
 
 	@Test
 	public void testSemver100_reduce_method_visibility() {
-		String lastPackage = "semver100.c";
-
-		String string = getSemverDiff(lastPackage);
-
-		assertEquals("1.0.0", string);
+		assertSemverDiffByPackage(MAJOR, "changesMethodFromPublicToDefault");
 	}
 
 	@Test
 	public void testSemver100_superclass_with_field() {
-		String lastPackage = "semver100.d";
-
-		String string = getSemverDiff(lastPackage);
-
-		assertEquals("1.0.0", string);
+		assertSemverDiffByPackage(MAJOR, "superclassWithField");
 	}
 
-	private String getSemverDiff(String lastPackage) {
-		JarArchiveComparator jarArchiveComparator = newComparator(lastPackage);
+	private void assertSemverDiffByPackage(String expectedSemverDiff, String lastPackageName) {
+		assertEquals(expectedSemverDiff, getSemverDiffByPackage(lastPackageName));
+	}
+
+	private String getSemverDiffByPackage(String lastPackage) {
 		File oldFile = Helper.getArchive("japicmp-test-v1.jar");
 		File newFile = Helper.getArchive("japicmp-test-v2.jar");
 
-		List<JApiClass> jApiClasses = jarArchiveComparator.compare(oldFile, newFile);
+		List<JApiClass> jApiClasses = newComparatorFor(lastPackage).compare(oldFile, newFile);
 		Options options = new Options();
 		options.setNewArchive(newFile);
 		options.setOldArchive(oldFile);
@@ -89,9 +73,9 @@ public class SemverOutIntegTest {
 		return new SemverOut(options, jApiClasses).value();
 	}
 
-	private JarArchiveComparator newComparator(String lastPackage) {
+	private JarArchiveComparator newComparatorFor(String lastPackage) {
 		JarArchiveComparatorOptions options = new JarArchiveComparatorOptions();
-		String packageName = "japicmp.test." + lastPackage;
+		String packageName = "japicmp.test.semver." + lastPackage;
 		options.getPackagesInclude().add(new PackageFilter(packageName));
 		return new JarArchiveComparator(options);
 	}
